alerts added & fixed styles

+++ src/components/WishList/WishList.jsx
@@ -4,6 +4,8 @@ import { getFavorites, removeFavorite } from "../../redux/actions";
 import w from "../WishList/WishList.module.css";
 import { Link } from "react-router-dom";
 import { HiHeart } from "react-icons/hi"
+import Swal from "sweetalert2";
+
 
 function WishList() {
   const dispatch = useDispatch();
@@ -20,9 +22,27 @@ function WishList() {
   let favs = favorites["Favorites"]
 
   const handleDeleteFavorite = async(productId) => {
-    await dispatch(removeFavorite(userId, productId))
-    dispatch(getFavorites(userId))
-    }
+  Swal.fire({
+  title: 'Removing from wishlist',
+  text: "Do you want to delete this product from your wishlist?",
+  icon: 'warning',
+  showCancelButton: true,
+  confirmButtonColor: '#3085d6',
+  cancelButtonColor: '#d33',
+  confirmButtonText: 'Yes, delete it!'
+}).then((result) => {
+  if (result.isConfirmed) {
+    dispatch(removeFavorite(userId, productId))
+    Swal.fire(
+      'Deleted!',
+      'The product has been deleted.',
+      'success'
+      )
+      dispatch(getFavorites(userId))
+  }})
+}
+
+
  if(favs === undefined)
   return (
     <div className={w.wCont}>
@@ -33,7 +53,7 @@ if(userId)
   return(
     <div className={w.wCont}> 
     <h1> My wishlist</h1>
-      {Array.isArray(favs)? favs.map(el => (<div className={w.wItem} key={el.id}>
+      {Array.isArray(favs) && favs.length? favs.map(el => (<div className={w.wItem} key={el.id}>
          <img src={el.image} height='120px' width='auto' alt={el.name} />
          <div className={w.productName}>
           <h3>{el.name}</h3>
